% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{asesvis}
\alias{asesvis}
\alias{asesvis-package}
\title{asesvis: An R Visualization Toolkit for ASES}
\description{
An analytics toolkit comprising of a series of visualization functions. The toolkit has
a lot of featured funcionalities specifically designed for ASES.
}
\details{
The package is comprised of \describe{
\item{Palette}{\code{\link{hex_colors}()}, \code{\link{named_pal}()}}
\item{Plots}{\code{\link{chartBox}()}, \code{\link{chartColumn}()},
  \code{\link{chartLine}()}, \code{\link{chartTornado}()}}
}
}
\seealso{
\pkg{\link{aseskit}}
}
\author{
\strong{Maintainer}: Yiying Wang, \email{wangy@aetna.com}
}
\keyword{internal}
