% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box.R
\name{chartBox}
\alias{chartBox}
\title{Box Chart Based on \pkg{ggplot2}}
\usage{
chartBox(data, xvar, yvar, gvar = NULL, horizontal = FALSE,
  xlab = xvar, ylab = yvar, title = paste("Chart: boxplot of", yvar,
  "by", xvar), palette = getOption("init.pal"), alt.font = "SimHei",
  theme = "wsj", plot.background = "antiquewhite1", ...)
}
\arguments{
\item{data}{Data frame for ploting.}

\item{xvar}{Variable name of x.}

\item{yvar}{Variable name of y.}

\item{gvar}{Variable name of grouping factor.}

\item{horizontal}{Logical. If TRUE, the boxes are horizontally presented.
Default FALSE.}

\item{xlab}{Label wording of x axis. Default \code{xvar}.}

\item{ylab}{Label wording of y axis. Default \code{yvar}.}

\item{title}{Chart title. Default \code{paste("Chart: boxplot of", yvar," by", xvar)}.}

\item{palette}{Palette name, default 'aetnagreen'}

\item{alt.font}{Define alternative font name for plotting. Default 'SimHei'.}

\item{theme}{The theme name. Only accepts themes in \pkg{ggplot2} and \pkg{ggthemes}.
Default 'wsj'.}

\item{plot.background}{The color name of the plot background.
Default 'antiquewhite1'.}

\item{...}{ignore}
}
\value{
A ggplot2 object.
}
\description{
Box chart using ggplot2.
}
\examples{

}
\seealso{
\code{\link{named_pal}}, \code{\link{ggplot2}}, \code{\link{ggthemes}}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
