% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_column.R
\name{chartColumn}
\alias{chartColumn}
\title{Column Chart Based on \pkg{ggplot2}}
\usage{
chartColumn(data, xvar, yvar, gvar = NULL, title = paste("Chart:",
  yvar, "~", xvar), position = "dodge", annotate = "number2",
  xlab = xvar, ylab = yvar, palette = getOption("init.pal"),
  hline = data.frame(label = NA, linetype = "solid", value = NA),
  vline = data.frame(label = NA, linetype = "solid", value = NA),
  line = list(list(label = NA, linetype = "solid", arrow = FALSE, value =
  data.frame(x = NA, y = NA))), alt.font = "SimHei", theme = "wsj",
  plot.background = "antiquewhite1", ...)
}
\arguments{
\item{data}{Data frame for ploting.}

\item{xvar}{Variable name of x.}

\item{yvar}{Variable name of y.}

\item{gvar}{Variable name of grouping factor.}

\item{title}{String. Chart title. Default \code{paste("Chart:", yvar,"~", xvar)}}

\item{position}{'dodge' or 'stack'.}

\item{annotate}{Format of the annotation. 'number(n)', 'percent(n)', 'comma'
or 'dollar'.}

\item{xlab}{Label wording of x axis. Default \code{xvar}.}

\item{ylab}{Label wording of y axis. Default \code{yvar}.}

\item{palette}{Palette name, default 'aetnagreen'}

\item{hline}{A data frame with columns of 'label', 'linetype' and 'value'.
Each record makes a horizontal line.}

\item{vline}{A data frame with columns of 'label', 'linetype' and 'value'.
Each record makes a vertical line.}

\item{line}{A list. Default value is \code{list(list(label=NA, linetype='solid',
arrow=FALSE, value=data.frame(x=NA, y=NA)))}. You should define the label,
linetype, arrow and value (a data frame consisting of x and y) of the
free-style line.}

\item{alt.font}{Define alternative font name for plotting. Default 'SimHei'.}

\item{theme}{The theme name. Only accepts themes in \pkg{ggplot2} and \pkg{ggthemes}.
Default 'wsj'.}

\item{plot.background}{The color name of the plot background.
Default 'antiquewhite1'.}

\item{...}{ignore}
}
\value{
A ggplot2 object.
}
\description{
Columns (vertical bar) chart using ggplot2, with or without benchmarking lines.
}
\examples{
\dontrun{
}
}
\seealso{
\code{\link{named_pal}}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
