% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_line.R
\name{chartLine}
\alias{chartLine}
\title{Line Chart Based on \pkg{ggplot2}}
\usage{
chartLine(data, xvar, yvar, gvar = NULL, title = paste("Chart:", yvar,
  "~", xvar), annotate = NULL, xlab = NULL, ylab = NULL,
  palette = .pkgPara()$init.pal, alt.font = "SimHei", theme = "wsj",
  plot.background = "antiquewhite1", ...)
}
\arguments{
\item{data}{Data frame for ploting.}

\item{xvar}{Variable name of x.}

\item{yvar}{Variable name of y.}

\item{gvar}{Variable name of grouping factor.}

\item{title}{Chart title. Default \code{paste("Chart:",yvar,"~",xvar)}.}

\item{annotate}{Format of the annotation. 'number(n)', 'percent(n)', 'comma'
or 'dollar'.}

\item{xlab}{Label wording of x axis. Default \code{xvar}.}

\item{ylab}{Label wording of y axis. Default \code{yvar}.}

\item{palette}{Palette name, default 'aetnagreen'}

\item{alt.font}{Define alternative font name for plotting. Default 'SimHei'.}

\item{theme}{The theme name. Only accepts themes in \pkg{ggplot2} and \pkg{ggthemes}.
Default 'wsj'.}

\item{plot.background}{The color name of the plot background.
Default 'antiquewhite1'.}

\item{...}{ignore}
}
\value{
A ggplot2 object.
}
\description{
Line chart using ggplot2.
}
\examples{

}
\seealso{
\code{\link{getHexPal}}, \code{\link{ggplot2}}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
