% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data:palettes.info}
\alias{data:palettes.info}
\title{Dataset: Palettes info}
\description{
Palettes information including name, palette full name, class, maxcolors, 
category and colorblind.
}
\details{
It is a 171 * 5 data.frame: \cr
\describe{
  \item{palette}{character, the formal name of the palettes}
  \item{class}{character, the class name of the palette: "brewer", "ggtableau",
   "ggthemes", "ggsolarized", "viridis", "aetna", "grdevice"}
  \item{maxcolors}{numeric, the max number of colors for each palette}
  \item{category}{chracter, category of 'div(verging)', 'qual(itative)' or 
   'seq(uential)'}
  \item{colorblind}{logical, whether the palette is colorblind-friendly}
}
}
\examples{
\dontrun{
data(palette.info)
}
}
\references{
The data set is used for \code{\link{getPalette}} and \code{\link{getHexPal}}
}
\seealso{
\pkg{\link{RColorBrewer}}  \pkg{\link{ggthemes}}  \pkg{\link{grDevices}}  
\pkg{\link{viridis}}
}
\author{
Yiying Wang \email{wangy@aetna.com}
}
\keyword{data}
\keyword{palettes}
