% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{getHexPal}
\alias{getHexPal}
\alias{hex_colors}
\title{Generalized color values parser}
\usage{
getHexPal(palette, unlist = TRUE, ...)

hex_colors(palette, unlist = TRUE, ...)
}
\arguments{
\item{palette}{A palette name with/without number of colors,
or a vector of color names, or a hex color value. \itemize{
 \item \code{\link{col2rgb}}-compatible arguments, either color names, hex color
 strings, or positive integer for \code{\link{palette}}()[i]. \cr
 \item \code{\link{named_pal}}-compatible character arguments, in a
 \code{"palette(n)"} format.
}}

\item{unlist}{Logical, whether unlist the result. Defalt TRUE.}

\item{...}{Other arguments to pass to the function, e.g., alpha, theme, ....}
}
\value{
\itemize{
\item palette is col2rgb-compatible: a list of hex colors (lower case). \cr
\item palette is named_pal-compatible: a list of hex colors (lower case) (when
 in '\code{palette_name(3)}' format) or a list of functions (when in
 '\code{palette_name()}' format).
 }
Note: If \code{unlist}=TRUE, the function will try to unlist the output list.
}
\description{
Get colors from a named palette or other formats
}
\examples{
\dontrun{
## Get a hex vector based on col2rgb-compatible palette
getHexPal(c("red", "yellow"))  # return
# [1] "#ff0000" "#ffff00"
getHexPal("red", alpha=200)  # feed alpha to ... with rgb()
# [1] "#ff0000c8"
hex_colors(2L, alpha=200)  # returns palette()[2L] with alpha=200
# [1] "#ff0000c8"

## Get a hex vector based on a hex value
getHexPal("#fff")  # return
# [1] "#ffffff"

## Get a hex vector based on named_pal-compatible palette
getHexPal("terrain(4)")  # return
# [1] "#00a600" "#e6e600" "#ecb176" "#f2f2f2"
## equivalent to
hex_colors("terrain", n=4)  # or the currying form
hex_colors("terrian")(4)

getHexPal(c("terrain(2)", "magma(2)"), unlist=FALSE)  # returns
# [[1]]
# [1] "#00a600" "#f2f2f2"
#
# [[2]]
# [1] "#000004" "#fcfdbf"

getHexPal("terrain")  # returns the function named_pal("terrain")
}
}
\seealso{
\code{\link{getPalette}}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
