% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{getPalette}
\alias{getPalette}
\alias{named_pal}
\title{Hex color vector in a palette}
\usage{
getPalette(palette, n = NULL, show = FALSE, mode = c("na.fill",
  "no.fill", "repeat.fill"), ...)

named_pal(palette, n = NULL, show = FALSE, mode = c("na.fill",
  "no.fill", "repeat.fill"), ...)
}
\arguments{
\item{palette}{Character, the palette name (case insensitive).}

\item{n}{Numeric, the length of color vector. Default NULL, which indicates that
the function will return a function to generate color vectors.}

\item{show}{Logical, whether display the palette. Default FALSE. If \code{n} is
NULL, then \code{show} will be ignored.}

\item{mode}{Character, either "na.fill", "no.fill", or "repeat.fill". If \code{n} >
max number of colors of the palette, then \itemize{
\item na.fill: Default choice. NAs will be used to fill the output vector \cr
\item no.fill: the output color vector will not be suplemented \cr
\item repeat.fill: the color vector will be repeated to fill the length
}}

\item{...}{Other arguments to pass to the function. E.g., \code{theme='Atlas'}
can be passed to \code{\link[ggthemes]{excel_new_pal}()} when \code{palette='excel_new'};
or \code{direction = -1} can be passed to \code{\link[ggthemes]{tableau_color_pal}}
when \code{palette='tableau10'}. You can always set \code{alpha} to change the
transparency of the output vector. \itemize{
 \item if \code{alpha} is between [0, 1] and is double, the [0, 1] alpha scale
 will be applied; \cr
 \item if \code{alpha} is integer (e.g., 2L, 155L, ...), the [0, 255] alpha
 scale will be applied.}}
}
\value{
\itemize{
\item \code{n} is not NULL: return a vector of Hex colors \cr
\item \code{n} is NULL: return the color generation function (function factory).}
}
\description{
Get a vector of hex colors (lower case characters) or a hex color function
factory from a designated (named) palette.
}
\details{
\code{getPalette} (\code{named_pal}) generally gets hex value of the
colors in a designated palette. The palettes include Aetna palette, palette
in packages \pkg{RColorBrewer}, \pkg{ggthemes} and \pkg{viridis}.
You can run \code{palettes.info} to check the
palettes information. \cr The following palettes are available:
\describe{
 \item{\strong{Aetna palttes}}{\describe{
   \item{'qual'}{'aetclassic_green', 'aetclassic_blue', 'aetclassic_teal',
    'aetclassic_cranberry', 'aetclassic_orange', 'aetclassic_violet',
    'aetnew_violet', 'aetnew_teal', 'aetnew_cranberry'}
   \item{'div'}{'aetdiv_violet', 'aetdiv_teal', 'aetdiv_cranberry'}
   \item{'seq'}{'aetdiv_violet', 'aetdiv_teal', 'aetdiv_cranberry'}}}
 \item{\strong{\pkg{\link[RColorBrewer]{RColorBrewer}} palettes}}{\describe{
   \item{'div'}{'BrBG', 'PiYG', 'PRGn', 'PuOr', 'RdBu', 'RdGy', 'RdYlBu',
    'RdYlGn', 'Spectral'}
   \item{'qual'}{'Accent', 'Dark2', 'Paired', 'Pastel1', 'Pastel2', 'Set1',
    'Set2', 'Set3'}
   \item{'seq'}{'Blues', 'BuGn', 'BuPu', 'GnBu', 'Greens', 'Greys', 'Oranges',
    'OrRd', 'PuBu', 'PuBuGn', 'PuRd', 'Purples', 'RdPu', 'Reds', 'YlGn',
    'YlGnBu', 'YlOrBr', 'YlOrRd'}}}
 \item{\strong{\pkg{\link{ggthemes}} palettes}}{\itemize{
   \item normal: \describe{
    \item{'qual'}{'calc', 'economist', 'economist_white', 'excel_line', 'excel_fill',
    'excel_new', 'few', 'fivethirtyeight', 'gdocs', 'pander', 'stata', 'stata1',
    'stata1r','statamono','hc', 'darkunica','wsj', 'wsj_rgby', 'wsj_red_green',
    'wsj_black_green', 'wsj_dem_rep', 'colorblind'}} \cr
   \item tableau (see \code{\link[ggthemes]{tableau_color_pal}}): \describe{
     \item{'regular' ('qual')}{"tableau10", "tableau20", "tbl_colorblind",
    "tbl_seattle_grays", "tbl_traffic", "tbl_miller_stone", "tbl_superfishel_stone",
    "tbl_nuriel_stone", "tbl_jewel_bright", "tbl_summer", "tbl_winter", "tbl_gnorte",
    "tbl_rdbubw", "tbl_pupkgy", "tbl_hue_circle", "tbl_classic10", "tbl_classic10_medium",
    "tbl_classic10_light", "tbl_classic20", "tbl_classic_gray5", "tbl_classic_colorblind",
    "tbl_classic_traffic_light", "tbl_classic_pugy6", "tbl_classic_pugy12",
    "tbl_classic_gnor6", "tbl_classic_gnor12", "tbl_classic_burd6", "tbl_classic_burd12",
    "tbl_classic_cyclic"}
    \item{'ordered-diverging' ('div')}{"tbl_orbu_div", "tbl_rdgn_div", "tbl_gnbu_div",
    "tbl_rdbu_div", "tbl_rdbl_div", "tbl_gdpu_div", "tbl_rdgngd_div",
    "tbl_sunset_sunrise_div", "tbl_orbuwi_div", "tbl_rdgnwi_div", "tbl_gnbuwi_div",
    "tbl_rdbuwi_div", "tbl_rdblwi_div", "tbl_orbu_light_div", "tbl_temperature_div",
    "tbl_classic_rdgn", "tbl_classic_rdbu", "tbl_classic_rdbl", "tbl_classic_area_rdgn",
    "tbl_classic_orbu", "tbl_classic_gnbu", "tbl_classic_rdwign", "tbl_classic_rdwibl",
    "tbl_classic_orwibu", "tbl_classic_rdwibl_light", "tbl_classic_orwibu_light",
    "tbl_classic_rdwign_light", "tbl_classic_rdgn_light"}
    \item{'ordered-sequential' ('seq')}{"tbl_bugn_seq", "tbl_blue_light", "tbl_orange_light",
    "tbl_blue", "tbl_orange", tbl_green", "tbl_red", "tbl_purple", "tbl_brown",
    "tbl_gray", "tbl_gray_warm", "tbl_bute", "tbl_orgd", "tbl_gngd", "tbl_rdgd",
    "tbl_classic_green", "tbl_classic_gray", "tbl_classic_blue", "tbl_classic_red",
    "tbl_classic_orange", "tbl_classic_area_red", "tbl_classic_area_green",
    "tbl_classic_area_brown"}
  } \cr
  \item solarized (see \code{\link[ggthemes]{solarized_pal}}): \describe{
   \item{'div'}{'solarized','solarized_red', 'solarized_yellow',
   'solarized_orange','solarized_magenta','solarized_violet',
   'solarized_blue', 'solarized_cyan', 'solarized_green'}}}}
 \item{\strong{\pkg{\link{viridis}} palettes}}{\describe{
  \item{'seq'}{'megma', 'inferno', 'plasma', 'viridis', 'cividis'}}}
 \item{\strong{\pkg{\link{grDevices}} palettes}}{\describe{
  \item{'qual'}{'rainbow'}
  \item{'seq'}{'terrain', 'topo', 'heat'}
  \item{'div'}{'cm'}}}
}
}
\examples{
\dontrun{
## Get the entire Aetna classic green palette
f <- getPalette("aetclassic_gre")  # returns a function factory
f(8)  # returns first 8 colors

## Get 7 colors from palette 'terrain'
named_pal("terrain", 7)
}
}
\seealso{
The palettes majorly come from \pkg{\link{RColorBrewer}}, \pkg{\link{ggthemes}},
and \pkg{\link{viridis}}. You can also refer to \pkg{\link{colorspace}} to build
your own palette. \cr
Refer to \code{\link{data:palettes.info}} to check built palettes info. And you
can use the generalized parser function \code{\link{hex_colors}} to get hex
colors more conveniently.
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
